Conflicts:
	apps/CecogAnalyzer/CecogAnalyzer.py
	apps/CecogAnalyzer/setup.py
	pysrc/cecog/__init__.py
	pysrc/cecog/analyzer/analyzer.py
	pysrc/cecog/analyzer/celltracker.py
	pysrc/cecog/analyzer/core.py
	pysrc/cecog/gui/modules/display.py
	pysrc/cecog/util/util.py

+++ csrc/include/cecog/morpho/geodesy.hxx
@@ -203,7 +203,7 @@ namespace morpho {
     int maxLabel = resStack.size();
     int numberOfMinMax = maxLabel;
 
-    if(maxLabel > pow(2, 8* sizeof(typename Accessor2::value_type)) - 1)
+    if(maxLabel > pow(2.0, (int) (8* sizeof(typename Accessor2::value_type)) - 1))
     {
       cout << "warning in ImMinMaxLabel: unappropriate output image value type." << endl;
       maxLabel = 1;
